% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDLData.R
\name{GetCDLData}
\alias{GetCDLData}
\title{Request for the CDL raster data}
\usage{
GetCDLData(aoi = NULL, year = NULL, type = NULL, mat = FALSE, crs = NULL)
}
\arguments{
\item{aoi}{Area of interest. Could be a 5-digit FIPS code of a county, three coordinates that defines a triangle,
or four corner points that defines a rectangle, or a single coordinate. The default coordinate system (used by CDL) is the Albers equal-area conic projection, or Albers projection. Users can provide
coordinates from a different projection method, but user have to specify the coordinate system in the \code{crs} argument.
For example, users can provide longitude/latitude coordinates here, while letting \code{crs} be '+init=epsg:4326'.}

\item{year}{Year of data to request. Can be a numerical value or a character.}

\item{type}{Type of aoi. 'f' for county, 'ps' for points, 'b' for box, 'p' for a single point.}

\item{mat}{TRUE/FALSE. If TRUE, return a data table. If FALSE (default), return a raster file.}

\item{crs}{Coordinate system, such as '+init=epsg:4326' for longitude/latitude.}
}
\value{
The function returns a raster file or a data table that saves the land use information. The file that matches the category codes with the category names
are available at~\url{https://www.nass.usda.gov/Research_and_Science/Cropland/docs/cdl_codes_names.xlsx}. Or simply use data(linkdata)
}
\description{
A function that makes HTTP GET requests for CDL raster data for any area of interests in a given crop year.
This function implements the GetCDLStat service provided by the CropScape~\url{https://nassgeodata.gmu.edu/CropScape}.
}
\examples{

# Example 1. Retrieve data for the Champaign county in Illinois (FIPS = 17109) in 2018.
data <- GetCDLData(aoi = 17019, year = 2018, type = 'f')
raster::plot(data) # plot the data.

# Example 2. Retrieve data for a single point by long/lat in 2018.
data <- GetCDLData(aoi = c(-94.6754,42.1197), year = 2018, type = 'p', crs = '+init=epsg:4326')
data
# Below uses the same point, but under the default coordinate system
data <- GetCDLData(aoi = c(108777,2125055), year = 2018, type = 'p')
data

# Example 3. Retrieve data for a triangle defined by three coordinates in 2018.
data <- GetCDLData(aoi = c(175207,2219600,175207,2235525,213693,2219600), year = 2018, type = 'ps')
raster::plot(data)

# Example 4. Retrieve data for a rectangle box defined by three corner points in 2018.
data <- GetCDLData(aoi = c(130783,2203171,153923,2217961), year = '2018', type = 'b')
raster::plot(data)

}
