% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkdata.R
\docType{data}
\name{linkdata}
\alias{linkdata}
\title{A dataset on adoption of conservation agriculture in Zambia}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 256 rows and 2 columns.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/docs/cdl_codes_names.xlsx}
}
\usage{
data(linkdata)
}
\description{
A dataset on adoption of conservation agriculture in Zambia
}
\keyword{datasets}
