% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDLComp.R
\name{GetCDLComp}
\alias{GetCDLComp}
\title{Request for comparisons of the CDL raster data}
\usage{
GetCDLComp(aoi, year1, year2, type, mat = TRUE, crs = NULL)
}
\arguments{
\item{aoi}{Area of interest. Could be a 5-digit FIPS code of a county, three coordinates that defines a triangle,
or four corner points that defines a rectangle. The default coordinate system (used by CDL) is the Albers equal-area conic projection, or Albers projection. Users can provide
coordinates from a different projection method, but user have to specify the coordinate system in the \code{crs} argument.
For example, users can provide longitude/latitude coordinates here, while letting \code{crs} be '+init=epsg:4326'.}

\item{year1}{Can be a numerical value or a character.}

\item{year2}{Can be a numerical value or a character.}

\item{type}{Type of aoi. 'f' for county, 'ps' for points, 'b' for box.}

\item{mat}{TRUE/FALSE. If TRUE (default), return a data table. If FALSE, return a raster file;}

\item{crs}{Coordinate system, such as '+init=epsg:4326' for longitude/latitude.}
}
\value{
The function returns a data table or a raster file.
}
\description{
A function that makes HTTP GET requests for cropland change analysis bewteen two years (year1 to year2) of the CDL raster data for an area of interests between two crop years.
This function implements the GetCDLComp service provided by the CropScape~\url{https://nassgeodata.gmu.edu/CropScape}.
}
\examples{

# Example 1. Retrieve data for the Champaign county in Illinois (FIPS = 17109) in 2017-2018.
data <- GetCDLComp(aoi = '17019', year1 = 2017, year2 = 2018, type = 'f')
head(data, 5)

# Example 2. Retrieve data for a triangle defined by three coordinates in 2017-2018.
aoi <- c(175207,2219600,175207,2235525,213693,2219600)
data <- GetCDLComp(aoi = aoi, year1 = 2017, year2 = 2018, type = 'ps')
head(data, 5)

# Example 3. Retrieve data for a rectangle box defined by four corner points in 2018.
data <- GetCDLComp(aoi = c(130783,2203171,153923,2217961), year1 = 2017, year2 = 2018, type = 'b')
head(data, 5)

}
