% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDLImage.R
\name{GetCDLImage}
\alias{GetCDLImage}
\title{Request for images of the CDL raster data}
\usage{
GetCDLImage(
  aoi = NULL,
  year = NULL,
  type = NULL,
  format = "png",
  crs = NULL,
  destfile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{aoi}{Area of interest. Can be a 5-digit FIPS code of a county, three coordinates that defines a triangle,
or four corner points that defines a rectangle. The default coordinate system (used by CDL) is the Albers equal-area conic projection, or Albers projection. Users can provide
coordinates from a different projection method, but user have to specify the coordinate system in the \code{crs} argument.
For example, users can provide longitude/latitude coordinates here, while letting \code{crs} be '+init=epsg:4326'.}

\item{year}{Year of data to request. Can be a numerical value or a character.}

\item{type}{Type of aoi. 'f' for county, 'ps' for points, 'b' for box.}

\item{format}{Format of the image file. Can be png or kml.}

\item{crs}{Coordinate system, such as '+init=epsg:4326' for longitude/latitude.}

\item{destfile}{A character string (or vector, see \code{url}) with the name where the downloaded file is saved. If not specified, the function will creater a temporary folder to save the image file.}

\item{verbose}{TRUE/FALSE. Display the directory saving the file or not.}
}
\value{
The function downloads an image file to users' computer.
}
\description{
A function that makes HTTP GET requests for the CDL raster data for an area of interests in a given crop year.
This function implements the GetCDLImage service provided by the CropScape~\url{https://nassgeodata.gmu.edu/CropScape}.
}
